Move Parser class to command package
