+++ src/main/java/mummy/command/Command.java
@@ -6,7 +6,6 @@ import java.util.HashMap;
 import mummy.task.Task;
 import mummy.task.TaskList;
 import mummy.ui.MummyException;
-import mummy.utility.Parser;
 import mummy.utility.Storage;
 
 /**
