+++ src/main/java/mummy/command/DeleteCommand.java
@@ -6,7 +6,6 @@ import mummy.task.Task;
 import mummy.task.TaskList;
 import mummy.task.TaskListException;
 import mummy.ui.MummyException;
-import mummy.utility.Parser;
 import mummy.utility.Storage;
