+++ src/main/java/mummy/command/UnmarkCommand.java
@@ -5,7 +5,6 @@ import java.util.HashMap;
 import mummy.task.TaskList;
 import mummy.task.TaskListException;
 import mummy.ui.MummyException;
-import mummy.utility.Parser;
 import mummy.utility.Storage;
 
 /**
